requirements.txt
bcrypt>=4.0.0
flask>=3.0.0
werkzeug>=3.0.0
spacy>=3.7.0
transformers>=4.35.0
nltk>=3.8.0
pygame>=2.5.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
PyJWT>=2.8.0
requests>=2.31.0